# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Proxy Tejelanas Vivi",
        "description": "API proxy para consultar información de Tejelanas Vivi con CORS habilitado",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor proxy local"
        }
    ],
    "paths": {
...
        "/api/v1/workshops": {
            "get": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Obtener lista de talleres",
                "operationId": "021a79b046ed6c7464f3f6c6a349dc74",
                "responses": {
                    "200": {
                        "description": "Lista de talleres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Workshop"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Crear nuevo taller",
                "operationId": "36306da80cf57dbcdd8612a622e78b31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Crochet Básico"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-15"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "14:00"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "max_participants": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Taller creado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/workshops/{id}": {
            "get": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Obtener taller específico",
                "operationId": "b670675804aacd742ff5b376cdb4e093",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Taller obtenido"
                    }
                }
            },
            "put": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Actualizar taller",
                "operationId": "b6153f66b525efb16301602c02f5da7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Taller actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Workshops"
                ],
                "summary": "Eliminar taller",
                "operationId": "b1f8e7723e22e581cd1ebc3a32f4caa2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Taller eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Workshop": {
                "title": "Workshop",
                "description": "Taller de crochet de Tejelanas Vivi",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Taller de Crochet Básico"
                    },
                    "description": {
                        "type": "string",
                        "example": "Aprende los fundamentos del crochet"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15"
                    },
                    "time": {
                        "type": "string",
                        "format": "time",
                        "example": "14:00:00"
                    },
                    "duration": {
                        "type": "integer",
                        "example": 120
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 25000
                    },
                    "max_participants": {
                        "type": "integer",
                        "example": 8
                    },
                    "current_participants": {
                        "type": "integer",
                        "example": 3
                    },
                    "location": {
                        "type": "string",
                        "example": "TEJElANAS, Laguna de Zapallar"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "full",
                            "cancelled"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token de autenticación Bearer. Tokens válidos para pruebas: tejelanas_admin_token_2025, development_token_123, test_token_456"
            }
        }
    },
...
}